# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MINIMALL API
  description: |
    Ultra-fast link-in-bio platform API for Shopify merchants.
    
    ## Overview
    MINIMALL provides a comprehensive API for managing link-in-bio site configurations, 
    assets, analytics, and user interactions. The API is built with Next.js API routes
    and follows RESTful principles.
    
    ## Authentication
    Currently, the API uses shop-based authentication. Authentication improvements
    are planned for future releases.
    
    ## Rate Limiting
    API requests are subject to reasonable rate limits to ensure service stability.
    
    ## Error Handling
    All endpoints return consistent error responses with appropriate HTTP status codes
    and detailed error messages.
  version: 1.0.0
  contact:
    name: MINIMALL API Support
    email: api-support@minimall.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Admin API (Development)
  - url: http://localhost:3000
    description: Public API (Development)
  - url: https://admin.minimall.com
    description: Admin API (Production)
  - url: https://api.minimall.com
    description: Public API (Production)
paths:
...
  /api/configs/{configId}/publish:
    post:
      tags:
        - Configurations
      summary: Publish configuration
      description: Publish a draft configuration to production
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Configuration published successfully"
                  version:
                    type: string
                    example: "v1704067200000"
                  publishedAt:
                    type: string
                    format: date-time
        '404':
          description: No draft version found to publish
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Asset Management
...
  /api/analytics/data:
    get:
      tags:
        - Analytics
      summary: Get analytics data
      description: Retrieve analytics data for performance monitoring and insights
      parameters:
        - name: configId
          in: query
          description: Filter by configuration ID
          schema:
            type: string
        - name: startDate
          in: query
          description: Start date for data range
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for data range
          schema:
            type: string
            format: date-time
        - name: timeframe
          in: query
          description: Predefined timeframe
          schema:
            type: string
            enum: ['1h', '24h', '7d', '30d']
            default: '24h'
      responses:
        '200':
          description: Analytics data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsData'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '503':
          description: Database not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PerformanceMetric:
      type: object
      required:
        - metric
        - value
        - rating
        - delta
        - id
        - navigationType
        - timestamp
        - url
      properties:
        configId:
          type: string
          example: "demo-shop-config"
        metric:
          type: string
          enum: [LCP, FID, CLS, FCP, TTFB]
          example: "LCP"
        value:
          type: number
          example: 2500
        rating:
          type: string
          enum: [good, needs-improvement, poor]
          example: "good"
        delta:
          type: number
          example: 100
        id:
          type: string
          example: "v3-1704067200-123"
        navigationType:
          type: string
          example: "navigate"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        url:
          type: string
          format: uri
          example: "https://example.com/g/demo-shop"
        userAgent:
          type: string
          example: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
        connection:
          type: string
          example: "4g"
        viewport:
          type: object
          properties:
            width:
              type: integer
              example: 1920
            height:
              type: integer
              example: 1080
    AnalyticsEvent:
      type: object
      required:
        - event
        - sessionId
      properties:
        event:
          type: string
          example: "button_click"
        configId:
          type: string
          example: "demo-shop-config"
        userId:
          type: string
          example: "user_123"
        sessionId:
          type: string
          example: "session_1704067200"
        properties:
          type: object
          additionalProperties: true
          example: 
            button_text: "Shop Now"
            section: "hero"
        userAgent:
          type: string
          example: "Mozilla/5.0 (iPhone; CPU iPhone OS 15_0 like Mac OS X)"
        referrer:
          type: string
          format: uri
          example: "https://instagram.com"
        utmSource:
          type: string
          example: "instagram"
        utmMedium:
          type: string
          example: "social"
        utmCampaign:
          type: string
          example: "summer_sale"
        utmTerm:
          type: string
          example: "fashion"
        utmContent:
          type: string
          example: "story_link"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    AnalyticsData:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            timeframe:
              type: object
              properties:
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
                duration:
                  type: string
                  example: "24h"
            performance:
              type: object
              properties:
                metrics:
                  type: array
                  items:
                    $ref: '#/components/schemas/PerformanceMetric'
                aggregates:
                  type: object
                  properties:
                    avgLcp:
                      type: number
                      example: 2200
                    avgFid:
                      type: number
                      example: 85
                    avgCls:
                      type: number
                      example: 0.05
                    avgTtfb:
                      type: number
                      example: 450
                    totalMetrics:
                      type: integer
                      example: 150
            analytics:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnalyticsEvent'
                eventCounts:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    page_view: 450
                    button_click: 127
                    product_view: 89
                uniqueSessions:
                  type: integer
                  example: 245
                totalEvents:
                  type: integer
                  example: 666
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Configuration not found"
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid configuration data"
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: "invalid_type"
              expected:
                type: string
                example: "string"
              received:
                type: string
                example: "number"
              path:
                type: array
                items:
                  type: string
                example: ["settings", "shopDomain"]
              message:
                type: string
                example: "Expected string, received number"
...